requirements.txt
amqp==5.2.0
antlr4-python3-runtime==4.9.3
audiomatch==0.2.2
audioread==3.0.1
billiard==4.2.0
black==24.4.2
blinker==1.8.2
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.0.0
colorama==0.4.6
cryptography==42.0.7
decorator==5.1.1
demucs==4.0.1
dora_search==0.1.12
einops==0.8.0
fastdtw==0.3.4
filelock==3.14.0
Flask==3.0.3
Flask-Executor==1.0.0
fsspec==2024.6.0
greenlet==3.0.3
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
julius==0.2.7
kombu==5.3.7
lameenc==1.7.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.42.0
MarkupSafe==2.1.5
mccabe==0.7.0
mkl==2021.4.0
mpmath==1.3.0
msgpack==1.0.8
mypy-extensions==1.0.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
omegaconf==2.3.0
openunmix==1.3.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
pooch==1.8.1
prompt_toolkit==3.0.47
pycodestyle==2.11.1
pycparser==2.22
pydub==0.25.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python_speech_features==0.6
PyYAML==6.0.1
redis==5.0.5
requests==2.32.3
retrying==1.3.4
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
submitit==1.5.1
sympy==1.12.1
tbb==2021.12.0
threadpoolctl==3.5.0
torch==2.3.1
torchaudio==2.3.1
tqdm==4.66.4
treetable==0.2.5
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
